factor out config and respect log_level
